home_page on the ptica(heliohost) done.
Problem with the session variable found and fixed. However I didn't remove the comments. Passing of the session variable works on the heliohost too ( it worked in the Eclipse locally ). There is the problem with the showing of the HTML page when I click on the book on the index page but passing of the session variable works fine.